requirements.txt
aiogram==3.13.1
google-genai>=1.0.0,<2.0
edge-tts>=6.1.10
python-dotenv==1.0.1
aiosqlite==0.20.0
APScheduler==3.10.4
aiohttp==3.10.10
